fix points_kill warnings
--HG--
branch : jowie

+++ src/scoring/points_kill.c
@@ -75,7 +75,7 @@ local int MyKillPoints(Arena *arena, Player *killer, Player *killed, int bounty,
 	/* cfghelp: Kill:FlagMinimumBounty, arena, int, def: 0
 	 * The minimum bounty the killing player must have to get any bonus
 	 * kill points for flags transferred, carried or owned. */
-	if (killer->position.bounty >=
+	if ((int) killer->position.bounty >=
 	    cfg->GetInt(arena->cfg, "Kill", "FlagMinimumBounty", 0))
 	{
 		/* cfghelp: Kill:PointsPerKilledFlag, arena, int, def: 100
